Promote VeryExperimental ValueCombiners to stable
