+++ MvvmCross/Core/Binding/MvvmCross.Binding.csproj
@@ -64,15 +64,15 @@
     <Compile Include="Bindings\Source\Construction\MvxPropertySourceBindingFactoryExtension.cs" />
     <Compile Include="Bindings\Target\MvxConvertingTargetBinding.cs" />
     <Compile Include="Combiners\IMvxValueCombiner.cs" />
-    <Compile Include="Combiners\VeryExperimental\MvxBooleanValueCombiner.cs" />
-    <Compile Include="Combiners\VeryExperimental\MvxEqualToValueCombiner.cs" />
+    <Compile Include="Combiners\MvxBooleanValueCombiner.cs" />
+    <Compile Include="Combiners\MvxEqualToValueCombiner.cs" />
     <Compile Include="Combiners\MvxFormatValueCombiner.cs" />
-    <Compile Include="Combiners\VeryExperimental\MvxGreaterThanOrEqualToValueCombiner.cs" />
-    <Compile Include="Combiners\VeryExperimental\MvxGreaterThanValueCombiner.cs" />
+    <Compile Include="Combiners\MvxGreaterThanOrEqualToValueCombiner.cs" />
+    <Compile Include="Combiners\MvxGreaterThanValueCombiner.cs" />
     <Compile Include="Combiners\MvxIfValueCombiner.cs" />
-    <Compile Include="Combiners\VeryExperimental\MvxLessThanOrEqualToValueCombiner.cs" />
-    <Compile Include="Combiners\VeryExperimental\MvxLessThanValueCombiner.cs" />
-    <Compile Include="Combiners\VeryExperimental\MvxNotEqualToPairwiseValueCombiner.cs" />
+    <Compile Include="Combiners\MvxLessThanOrEqualToValueCombiner.cs" />
+    <Compile Include="Combiners\MvxLessThanValueCombiner.cs" />
+    <Compile Include="Combiners\MvxNotEqualToPairwiseValueCombiner.cs" />
     <Compile Include="Combiners\MvxSingleValueCombiner.cs" />
     <Compile Include="Bindings\SourceSteps\MvxSourceStep.cs" />
     <Compile Include="Bindings\SourceSteps\MvxSourceStepDescription.cs" />
